refactor: extract contact message
